+++ packages/provider/src/enums/index.ts
@@ -1,3 +1,1 @@
-export * from './near-actions'
 export * from './providers'
-export * from './solana-rpc-error-codes'