+++ packages/provider/src/enums/near-actions.ts
@@ -1,3 +1,0 @@
-export enum NEAR_ACTIONS {
-  functionCall = 'FunctionCall',
-}