+++ packages/providers-solana/src/base-solana.ts
@@ -5,27 +5,27 @@ import {
   ChainTypes,
   SOLANA_CHAIN_IDS,
 } from '@rarimo/shared'
-import { PublicKey } from '@solana/web3.js'
 
-import { ProviderEventBusEvents, ProviderEvents } from '@/enums'
 import {
-  getSolExplorerAddressUrl,
-  getSolExplorerTxUrl,
-  handleEthError,
-  handleSolError,
-} from '@/helpers'
-import type {
-  EthProviderRpcError,
+  errors,
+  ProviderEventBus,
+  ProviderEventBusEvents,
+  ProviderEvents,
   ProviderProxy,
   RawProvider,
   SolanaProvider,
-  SolanaProviderRpcError,
   SolanaTransactionResponse,
+  TransactionRequestBody,
   TransactionResponse,
-  TxRequestBody,
-} from '@/types'
+} from '@rarimo/provider'
+import { PublicKey } from '@solana/web3.js'
 
-import { ProviderEventBus } from './event-bus'
+import {
+  getSolExplorerAddressUrl,
+  getSolExplorerTxUrl,
+  handleSolError,
+} from '@/helpers'
+import type { SolanaProviderRpcError } from '@/types'
 
 const getAddress = (publicKey: PublicKey | null): string => {
   return publicKey ? new PublicKey(publicKey).toBase58() : ''
@@ -39,7 +39,8 @@ export class BaseSolanaProvider
   #chainId?: ChainId
   #address?: string
 
-  constructor(provider: RawProvider) {
+  constructor(provider?: RawProvider) {
+    if (!provider) throw new errors.ProviderInjectedInstanceNotFoundError()
     super()
     this.#provider = provider as SolanaProvider
   }
@@ -84,7 +85,7 @@ export class BaseSolanaProvider
     try {
       await this.#provider.connect()
     } catch (error) {
-      handleEthError(error as EthProviderRpcError)
+      handleSolError(error as SolanaProviderRpcError)
     }
   }
 
@@ -104,7 +105,7 @@ export class BaseSolanaProvider
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    txRequestBody: TxRequestBody,
+    txRequestBody: TransactionRequestBody,
   ): Promise<TransactionResponse> {
     throw new TypeError('Method should be implemented in extender class')
   }