+++ packages/providers-solana/src/solflare.ts
@@ -1,32 +1,33 @@
 import {
+  ProviderProxy,
+  Providers,
+  RawProvider,
+  TransactionRequestBody,
+  TransactionResponse,
+} from '@rarimo/provider'
+import {
   Cluster,
   clusterApiUrl,
   Connection,
   Transaction as SolTransaction,
 } from '@solana/web3.js'
 
-import { Providers } from '@/enums'
 import { decodeSolanaTx, handleSolError } from '@/helpers'
-import type {
-  ProviderProxy,
-  RawProvider,
-  SolanaProviderRpcError,
-  TransactionResponse,
-  TxRequestBody,
-} from '@/types'
+import type { SolanaProviderRpcError } from '@/types'
 
 import { BaseSolanaProvider } from './base-solana'
 
 /**
- * @description Represents a Phantom wallet.
+ * @description Represents a Solflare wallet.
  *
  * @example
  * ```js
- * import { createProvider, PhantomProvider } from '@rarimo/provider'
+ * import { createProvider } from '@rarimo/provider'
+ * import { SolflareProvider } from '@rarimo/providers-solana
  *
- * const getPhantomWalletAddress = async () => {
- *   // Connect to the Phantom wallet in the browser using Web3.js, using the PhantomProvider interface to limit bundle size.
- *   const provider = await createProvider(PhantomProvider)
+ * const getSolflareWalletAddress = async () => {
+ *   // Connect to the Solflare wallet in the browser, using the SolflareProvider interface to limit bundle size.
+ *   const provider = await createProvider(SolflareProvider)
  *   await provider.connect()
  *
  *   // Get the address of the wallet
@@ -34,20 +35,20 @@ import { BaseSolanaProvider } from './base-solana'
  * }
  * ```
  */
-export class PhantomProvider
+export class SolflareProvider
   extends BaseSolanaProvider
   implements ProviderProxy
 {
-  constructor(provider: RawProvider) {
+  constructor(provider?: RawProvider) {
     super(provider)
   }
 
   static get providerType(): Providers {
-    return Providers.Phantom
+    return Providers.Solflare
   }
 
   async signAndSendTx(
-    txRequestBody: TxRequestBody,
+    txRequestBody: TransactionRequestBody,
   ): Promise<TransactionResponse> {
     try {
       const txBody =
@@ -55,10 +56,14 @@ export class PhantomProvider
           ? decodeSolanaTx(txRequestBody)
           : txRequestBody
 
+      const signedTx = await this.provider.signTransaction(
+        txBody as SolTransaction,
+      )
+
       const connection = new Connection(clusterApiUrl(this.chainId as Cluster))
 
-      const { signature } = await this.provider.signAndSendTransaction(
-        txBody as SolTransaction,
+      const signature = await connection.sendRawTransaction(
+        signedTx.serialize(),
       )
       await connection.confirmTransaction(signature)
       return signature