+++ packages/react-provider/src/hooks/useProvider.ts
@@ -1,25 +1,12 @@
 import {
   createProvider,
   CreateProviderOpts,
-  EthereumProvider,
-  EthTransactionResponse,
   IProvider,
-  NearProviderType,
-  NearTransactionResponse,
-  NearTxRequestBody,
   ProviderBase,
   ProviderEventPayload,
-  ProviderInstance,
-  ProviderProxy,
   ProviderProxyConstructor,
-  Providers,
-  ProvidersReturnType,
-  RawProvider,
-  SolanaProvider,
-  SolanaTransactionResponse,
+  TransactionRequestBody,
   TransactionResponse,
-  TxRequestBody,
-  Web3,
 } from '@rarimo/provider'
 import type { Chain, ChainId, ChainTypes } from '@rarimo/shared'
 import { useCallback, useEffect, useState } from 'react'
@@ -72,7 +59,7 @@ export const useProvider = (
     provider.switchChain(chainId)
 
   const signAndSendTx = async (
-    txRequestBody: TxRequestBody,
+    txRequestBody: TransactionRequestBody,
   ): Promise<TransactionResponse> =>
     provider.signAndSendTx?.(txRequestBody) ?? ''
 
@@ -144,29 +131,4 @@ export const useProvider = (
     getHashFromTx,
     getAddressUrl,
   }
-}
-
-export type {
-  Chain,
-  ChainId,
-  ChainTypes,
-  CreateProviderOpts,
-  EthereumProvider,
-  EthTransactionResponse,
-  IProvider,
-  NearProviderType,
-  NearTransactionResponse,
-  NearTxRequestBody,
-  ProviderEventPayload,
-  ProviderInstance,
-  ProviderProxy,
-  ProviderProxyConstructor,
-  Providers,
-  ProvidersReturnType,
-  RawProvider,
-  SolanaProvider,
-  SolanaTransactionResponse,
-  TransactionResponse,
-  TxRequestBody,
-  Web3,
 }