+++ packages/react-provider/vite.config.ts
@@ -38,7 +38,6 @@ export default defineConfig(({ mode }) => {
       dedupe: ['react'],
       alias: {
         '@': `${root}/`,
-        'near-api-js': 'near-api-js/dist/near-api-js.js',
       },
     },
   }