+++ CHANGELOG.md
@@ -21,6 +21,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `all` - Node resolution in the `package.json` files
 
+## [1.5.0-rc.9] - 2023-05-12
+### Added
+- `root`: `@rarimo/providers-evm` package
+- `root`: `@rarimo/providers-near` package
+- `root`: `@rarimo/providers-solana` package
+
+### Changed
+- `@rarimo/provider`: Wallet wrappers split to the several packages: `@rarimo/providers-evm`, `@rarimo/providers-near`, `@rarimo/providers-solana`
+
+### Fixed
+- `@rarimo/nft-checkout`: Uniswap V3 token list url changed to the new one to omit gateway issues
+
 ## [1.5.0-rc.8] - 2023-04-25
 ### Added
 - `all` - `browser` resolution path in the `package.json`
@@ -195,7 +207,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.7...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.9...HEAD
+[1.5.0-rc.9]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.8...1.5.0-rc.9
+[1.5.0-rc.8]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.7...1.5.0-rc.8
 [1.5.0-rc.7]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.6...1.5.0-rc.7
 [1.5.0-rc.6]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.5...1.5.0-rc.6
 [1.5.0-rc.5]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.4...1.5.0-rc.5